# ring: Fix pathological case when an entire zone leaves