@56quarters 2025-03-25
Comparison against [a commit from early January](https://github.com/grafana/dskit/tree/56quarters/ring-prev) which predates my changes in #632:

```
$ benchstat prev.txt current.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/dskit/ring
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                           │   prev.txt    │             current.txt             │
                           │    sec/op     │   sec/op     vs base                │
Ring_Get_OneZoneLeaving-16   8329.3µ ± 13%   114.6µ ± 5%  -98.62% (p=0.000 n=10)

                           │   prev.txt   │           current.txt            │
                           │     B/op     │     B/op      vs base            │
Ring_Get_OneZoneLeaving-16   10.27Ki ± 0%   10.27Ki ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                           │  prev.txt  │          current.txt           │
                           │ allocs/op  │ allocs/op   vs base            │
Ring_Get_OneZoneLeaving-16   12.00 ± 0%   12.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```